(codacy) fixes on several files
