+++ dbConnection.py
@@ -1,9 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from __future__ import absolute_import
-from builtins import str
-from builtins import range
-from builtins import object
 from qgis.core import QgsDataSourceUri
 from qgis.PyQt.QtCore import QSettings
 from qgis.PyQt.QtWidgets import QAction
