+++ pgRoutingLayer_utils.py
@@ -87,22 +87,6 @@ def refreshMapCanvas(mapCanvas):
 def logMessage(message, level=Qgis.Info):
     QgsMessageLog.logMessage(message, 'pgRouting Layer', level)
 
-def getNodeQuery(args, geomType):
-    ''' returns a sql query to get nodes from a geometry. '''
-    return sql.SQL("""
-        WITH node AS (
-            SELECT {id},
-                ST_X({geometry}) AS x,
-                ST_Y({geometry}) AS y,
-                {geometry}
-            FROM (
-                SELECT {source} AS id, ST_StartPoint({geom_t}) AS _geom
-                FROM {edge_table}
-                    UNION
-                SELECT {target}, ST_EndPoint({geom_t})
-                FROM {edge_table}
-                ) AS node
-       )""").format(**args)
 
 def getPgrVersion(con):
     ''' returns version of PostgreSQL database. '''
@@ -120,58 +104,3 @@ def getPgrVersion(con):
         return 0
     except SystemError as e:
         return 0
-
-def get_innerQuery(args):
-    return sql.SQL("""
-        SELECT {id} AS id,
-                {source} AS source,
-                {target} AS target,
-                {cost}::FLOAT AS cost,
-                {reverse_cost}::FLOAT AS reverse_cost
-            FROM {edge_table}
-            {where_clause}
-        """.replace("\\n", r"\n")).format(**args)
-
-def get_innerQueryXY(args):
-    return sql.SQL("""
-        SELECT {id} AS id,
-                {source} AS source,
-                {target} AS target,
-                {cost}::FLOAT AS cost,
-                {reverse_cost}::FLOAT AS reverse_cost,
-                {x1}::FLOAT AS x1,
-                {y1}::FLOAT AS y1,
-                {x2}::FLOAT AS x2,
-                {y2}::FLOAT AS y2
-            FROM {edge_table}
-            {where_clause}
-        """.replace("\\n", r"\n")).format(**args)
-
-def get_closestVertexInfo(args):
-    return  sql.SQL("""
-        WITH
-        near_source AS(SELECT {source},
-                ST_Distance(
-                    ST_StartPoint({geom_t}),
-                    ST_GeomFromText('POINT({x} {y})', {dbcanvas_srid})
-                ) AS dist,
-                ST_AsText(ST_StartPoint({geom_t})) AS point
-                FROM {edge_table}
-                WHERE  {geom_t} && {SBBOX} ORDER BY dist ASC LIMIT 1
-        ),
-        near_target AS(SELECT {target},
-                ST_Distance(
-                    ST_EndPoint({geom_t}),
-                    ST_GeomFromText('POINT({x} {y})', {dbcanvas_srid})
-                ) AS dist,
-                ST_AsText(ST_EndPoint({geom_t}))
-                FROM {edge_table}
-                WHERE  {geom_t} && {SBBOX} ORDER BY dist ASC LIMIT 1
-        ),
-        the_union AS (
-            SELECT * FROM near_source UNION SELECT * FROM near_target
-        )
-        SELECT {source}, dist, point
-        FROM the_union
-        ORDER BY dist ASC LIMIT 1
-        """).format(**args)
